admin_dashboard/admin/requirements.txt
streamlit>=1.36.0
requests>=2.32.0
pandas>=2.2.0
python-dotenv>=1.0.1
plotly>=5.22.0
openpyxl>=3.1.5
